requirements.txt
nltk==3.8.1
pandas==2.0.0
torch==2.0.1
scikit-learn==1.2.2
transformers==4.30.2
streamlit==1.25.0
python-dotenv==1.0.0
langchain==0.3
langchain_community
langchain-openai
faiss-cpu==1.7.4
openai==0.28.1
shutilwhich==1.1.0
edgartools==3.0.1
tqdm>=4.60.0
yfinance>=0.2.0
numpy>=1.21.0
tiktoken
